rename pytest-verifun to pytest-funparam
I've always hated the name verifun, and I found just the excuse to
change it: While there are no matches for "verifun" on PyPI, it asks if
you meant to search for "verfun". That project seems legit cool, and
completely irrelevant to this one.

At any rate, "verifun" was never a good name for this project. It
doesn't even include "param" in the name, which is the whole point of
this thing. So goodbye verifun, and hello funparam!

I should probably find another name for "verify_functions" too, but I
really need to get this out the door.